requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
schedule==1.2.0
python-dotenv==1.0.0
openai==1.3.0
sqlalchemy==2.0.21
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-telegram-bot==20.7
